Abstract class implementation for CAN hardware
